Revert "[testsuite] Split Obj-C foundation test"
This reverts commit 8416369862670bd4b93b1b5ca0a66a3483c815b4.
